+++ src/repositories/fine-payment/fine-payment.repository.ts
@@ -1,6 +1,12 @@
-import { count, desc, or, ilike } from "drizzle-orm";
+import { count, desc, or, ilike, sql, eq } from "drizzle-orm";
 import { db } from "../../db";
-import { finePayments } from "../../db/schema";
+import {
+  books,
+  finePayments,
+  librarians,
+  members,
+  transactions,
+} from "../../db/schema";
 import redisClient from "../../config/redis";
 
 export async function findFinePaymentWithPagination(
@@ -16,23 +22,53 @@ export async function findFinePaymentWithPagination(
   }
 
   const offset = (page - 1) * limit;
+
   const whereClause = search
     ? or(
-        ilike(finePayments.hargaDenda as any, `%${search}%`),
-        ilike(finePayments.totalDenda as any, `%${search}%`),
-        ilike(finePayments.metodePembayaran as any, `%${search}%`),
+        ilike(sql`CAST(${finePayments.hargaDenda} AS TEXT)`, `%${search}%`),
+        ilike(sql`CAST(${finePayments.totalDenda} AS TEXT)`, `%${search}%`),
+        ilike(
+          sql`CAST(${finePayments.metodePembayaran} AS TEXT)`,
+          `%${search}%`,
+        ),
+        ilike(members.nama, `%${search}%`),
+        ilike(librarians.nama, `%${search}%`),
+        ilike(transactions.kdTransaksi, `%${search}%`),
+        ilike(books.judul, `%${search}%`),
       )
     : undefined;
 
   const [data, countResult] = await Promise.all([
     db
-      .select()
+      .select({
+        id: finePayments.id,
+        hargaDenda: finePayments.hargaDenda,
+        totalDenda: finePayments.totalDenda,
+        tglBayar: finePayments.tglBayar,
+        metodePembayaran: finePayments.metodePembayaran,
+        createdAt: finePayments.createdAt,
+        namaPustakawan: librarians.nama,
+        namaAnggota: members.nama,
+        kdTransaksi: transactions.kdTransaksi,
+        judulBuku: books.judul,
+      })
       .from(finePayments)
+      .innerJoin(librarians, eq(finePayments.pustakawanId, librarians.id))
+      .innerJoin(members, eq(finePayments.anggotaId, members.id))
+      .innerJoin(transactions, eq(finePayments.transaksiId, transactions.id))
+      .innerJoin(books, eq(transactions.bukuId, books.id))
       .where(whereClause)
       .orderBy(desc(finePayments.createdAt))
       .limit(limit)
       .offset(offset),
-    db.select({ total: count() }).from(finePayments).where(whereClause),
+    db
+      .select({ total: count() })
+      .from(finePayments)
+      .innerJoin(librarians, eq(finePayments.pustakawanId, librarians.id))
+      .innerJoin(members, eq(finePayments.anggotaId, members.id))
+      .innerJoin(transactions, eq(finePayments.transaksiId, transactions.id))
+      .innerJoin(books, eq(transactions.bukuId, books.id))
+      .where(whereClause),
   ]);
 
   const totalRows = Number(countResult[0]?.total ?? 0);